refactor: Исключения внесены в отдельный пакет. Удалена неиспользовавшаяся зависимость. Добавлива аннотация @Slf4j и использование централизованную обработку ошибок
