+++ pom.xml
@@ -43,14 +43,6 @@
 			<version>1.18.20</version>
 			<scope>provided</scope>
 		</dependency>
-
-			<dependency>
-				<groupId>com.google.code.gson</groupId>
-				<artifactId>gson</artifactId>
-				<version>2.8.8</version>
-			</dependency>
-
-
 		<dependency>
 			<groupId>org.zalando</groupId>
 			<artifactId>logbook-spring-boot-starter</artifactId>
