+++ src/main/java/ru/yandex/practicum/filmorate/exception/ValidationException.java
@@ -1,13 +1,13 @@
-package ru.yandex.practicum.filmorate;
+package ru.yandex.practicum.filmorate.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class ValidationException extends IllegalArgumentException {
-    public ValidationException(String message) {
-    }
+
     public ValidationException() {
 
     }
+
 }
