+++ src/main/resources/application.properties
@@ -1,4 +1,2 @@
-logging.level.ru.yandex.practicum.filmorate.controller.UserController=info
-logging.level.ru.yandex.practicum.filmorate.controller.FilmController=info
 logging.level.org.zalando.logbook=info
