+++ src/test/java/ru/yandex/practicum/filmorate/controller/FilmControllerTest.java
@@ -3,11 +3,9 @@ package ru.yandex.practicum.filmorate.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.validation.BindingResult;
-import ru.yandex.practicum.filmorate.ValidationException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.io.IOException;
@@ -22,25 +20,12 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class FilmControllerTest {
-    FilmController filmController = new FilmController();
     HttpClient client = HttpClient.newHttpClient();
     ObjectMapper mapper = new ObjectMapper();
     URI url;
     Film film = new Film();
     @LocalServerPort
     private int port;
-
-    @Test
-    void checkingWorkValidationExceptionFilmController() {
-        defaultValues();
-        BindingResult bindingResult = Mockito.mock(BindingResult.class);
-        Mockito.when(bindingResult.hasFieldErrors()).thenReturn(true);
-        assertThrows(ValidationException.class, () -> filmController.addFilm(film, bindingResult));
-
-        Mockito.when(bindingResult.hasFieldErrors()).thenReturn(false);
-        Film factUser = filmController.addFilm(film, bindingResult);
-        assertEquals(film, factUser);
-    }
 
     @Test
     void filmFieldValidationPositive() throws ValidationException, IOException, InterruptedException {
